+++ src/Conversation/Conversation.spec.js
@@ -5,9 +5,22 @@ import { shallow, mount } from 'enzyme';
 import { expect } from 'chai';
 
 describe('<Conversation />', () => {
-  it('renders one p element', () => {
-    const wrapper = shallow(<Conversation />);
-    expect(wrapper.find('p')).to.have.length(1);
-    expect(wrapper.find('p').text()).to.equal('Hello World!');
+  it('renders a Messages element with no messages', () => {
+    const wrapper = shallow(<Conversation messages={[]} />);
+    expect(wrapper.find('Messages')).to.have.length(1);
+    expect(wrapper.find('Message')).to.have.length(0);
+  });
+
+  it('renders one Messages element with one message', () => {
+    const wrapper = shallow(<Conversation messages={[
+      {
+        message:'Paul',
+        from: 'left',
+        backColor: 'white',
+        duration: 200
+      }
+    ]} />);
+    expect(wrapper.find('Messages')).to.have.length(1);
+    expect(wrapper.find('Message')).to.have.length(0);
   });
 });