+++ stories/conversation.js
@@ -3,10 +3,41 @@ import Conversation from '../src/Conversation/Conversation';
 import { storiesOf, action } from '@kadira/storybook';
 
 storiesOf('Conversation', module)
-  .add('default view', () => {
-    return (
-      <div>
-        <Conversation name={'Paul'}/>
-      </div>
-    );
-  });
+.add('should display no messages', () => {
+  return (
+    <div>
+      <Conversation messages={[]}/>
+    </div>
+  );
+}).add('should display all messages', () => {
+  return (
+    <div>
+      <Conversation messages={[
+        {
+          message:'Paul',
+          from: 'left',
+          backColor: 'white',
+          duration: 200
+        },
+        {
+          message:'Conor',
+          from: 'left',
+          backColor: 'white',
+          duration: 200
+        },
+        {
+          message:'JimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJim',
+          from: 'left',
+          backColor: 'white',
+          duration: 200
+        },
+        {
+          message:'JimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJimJim',
+          from: 'right',
+          backColor: '#dcf8c6',
+          duration: 200
+        }
+      ]}/>
+    </div>
+  )
+});